Re-organize residential care package factories
